DatasetInfo updates
